MulDownloadImages.py

+++ MulDownloadImages.py
@@ -176,7 +176,7 @@ if __name__ == '__main__':
     pagesize = 10000;
     print getTotalNum()[0][0]
     print getTotalNum()[0][0] * 1.0 / pagesize
-    pages = math.ceil(getTotalNum()[0][0] * 1.0 / pagesize)
+    pages = int(math.ceil(getTotalNum()[0][0] * 1.0 / pagesize))
     print pages;
     # (status, output) = commands.getstatusoutput('ps aux|grep python|grep -v grep')
     # print status, output
